Fix realizing into multiple existing buffers
The natural way to realize into multiple existing buffers currently is
to put them in a temporary Realization:

f.realize({im1, im2});

This doesn't do precisely you want (it used to, before the Buffer
change). The Realization constructor makes a copy of those buffer
objects that it can own, so you'll lose any mutation of the metadata
(e.g. device allocations). It's also different to this:

f.realize(im1);

which takes im1 by reference and updates the metadata

This PR adds a new realization-like class that collects an array of
mutable references to existing buffers. It's a temporary type solely
used for passing groups of buffers to realize and infer_input_bounds. It
changes the meaning of

f.realize({im1, im2})

to take im1 and im2 by reference, the way it used to be.


Former-commit-id: 73ee0d68d2f0d9b0b36631cf18c70fc6979111f6