Fix openglcompute app
Made it use Halide::Buffer.

While doing this I discovered that the vector support in OpenGLCompute
codegen was broken, and made a dangerous assumption anyway (all vector
access is aligned), so I neutered it for now. It needs a pass that
actually detects if all buffer access is aligned dense vectors, and if
so, declares the buffer as an array of vectors.


Former-commit-id: 475808a7d314d21d0d74da7ea7b1333627573fb8