Disable early autovectorization of runtime
We compile the runtime modules to .ll for a "generic" target at -O3.
Apparently in some circumstances this autovectorizes to 128-bits. Not
all vector types are supported by all targets (hexagon), so this is bad,
and can cause assertion failures inside LLVM during instruction
selection. This PR turns this off.

The .ll gets a second chance to autovectorize later in codegen once the
module is assembled and has a concrete target.