Fix explicit device_malloc and copy_to_device in JIT-using code
This previously failed in mysterious ways. Copying to device in JIT code
before any pipeline had run on device didn't properly initialize device
state.

Even after a pipeline had run on device there was no way to feed
copy_to_device the same Target as used in a subsequent realize (which is
necessary to set the Debug flag), and there was no way to get it to
automatically pick a device api that matched realize's default.

There are still major unresolved problems with JIT and GPU. For example
if you say:

b.device_malloc(Target("host-cuda"));

Then b gets used in an OpenCL pipeline, the underlying device allocation
was freed and replaced with an OpenCL one, but it was freed in a copy of
b's buffer struct. So b now contains a dev field that's a dangling
pointer. We get a crash in b's destructor.

The root cause of these issues is the private copy of the user's Buffer
object that we make. I have a separate branch I'm working on that lets
Halide pipelines carry around references to user buffers instead of
making private copies of them like this.


Former-commit-id: 37843815fb9f155cad0ae8d817defd2b9d6c0daa