Add JITSharedRuntime::set_jit_externs()
At present there is no good, reliable way to override arbitrary runtime
functions when using the JIT. JITSharedRuntime ::set_default_handlers()
allows overriding of selected runtime methods, but extending for
multiple entrypoints (e.g. all of a GPU backend, to allow running on a
simulator) is tedious. This PR adds the ability to selectively override
arbitrary functions in JIT compilation mode, including runtime
functions.

(It is likely that this change could be used to reimplement or simply
replace set_default_handlers(); I invite comments on the wisdom of
that.)

(Note that this is built atop
https://github.com/halide/Halide/pull/1723, and cannot land before it
does)