Revise get_test_tmp_dir()
Putting it under /tmp/ causes race conditions in Travis (etc), as
multiple tests contend for the same dir. Instead:
— If the TEST_TMPDIR env var is defined, use it. (Bazel defines this
env var automatically.)
— Otherwise, use the current working directory (essentially reverting
to the same behavior as we used to have, but via absolute path rather
than relative).

Assert that paths must be absolute.

(Pushing out for review, but I’m going to test this locally on Windows
before committing.)


Former-commit-id: a403b536f09ddd89891fa28199ddbaf946b2ab5e