Remove Generator multitarget special-cases
The code path for Generators formerly went thru compile_multitarget()
only for cases where there were multitargets; this meant that the
contents of static libraries (and the patterns of filename
substitution) could be subtly different for single- vs multi-target,
making some bugs hard to track down (since some Generator tests use
single and some use multi). Now, the code uniformly goes thru the
multitarget case; this means that a ‘wrapper’ function is now generated
for the single-target case where none was before, but in terms of code
size and speed, it’s essentially a wash (the wrapper is just the an
unconditional branch to the underlying function). This will make
subsequent changes to this code much easier to verify across all usages.


Former-commit-id: dc172d9db22dcb31477968ef220481ae5c968a17