print test should tread "-nan" output as "nan"
AFAIK, “-nan” is not actually a thing in IEEE754, and the fact that
some printf implementations can emit this string for some bit patterns
is a bug. The test here was special-casing it for __APPLE__, but I am
seeing it in other builds too.

(Note the old comment was wrong, and should have read “OSX prints some
nan values as -nan”)