Convert IROperator.h args to const-ref where trivial
Convert Expr-by-value args to Expr-by-const-ref in IROperator.h; this
is motivated by reducing of temporaries (and thus stack space) for
debug/non-opt builds of Halide.

This is a very conservative change: only functions that could continue
to work unmodified were changed (if the args were mutated in any way,
pass-by-value is retained).

(It’s arguably preferable to be consistent within a file and convert
100% of the calls to const-ref, even if it means minor cleanup to the
interior code of a few methods; if reviewers prefer this be done here,
LMK)