Reduce overhead of internal_assert and user_assert
Use a time-honored operator-precedence trick to avoid evaluating any of
the stream output for these assertions when the assertion is true. On
my four-core laptop, this gives a pretty consistent reduction in the
fuzz_simplify test from ~6.0s -> 5.6s.

(Note that if any code was previously relying on side-effects from
assertion args always being evaluated, it will break; hopefully that’s
not the case…)

Note the two TODO comments in the code; both should be resolved before
possibly landing this, so comments welcome.


Former-commit-id: 332770715ba08255115c8af36ebd60629ef148f8