Fix camera_pipe viz script
There's still a minor problem with viz: sometimes Halide intentionally
loads out of bounds memory, and uses it to compute and store garbage
values, when it knows those values can't affect the output. These show
up as noise in the viz. This is true to what's going on, but looks
pretty distracting. The visualizer doesn't track bounds in a detailed
enough way to detect these things.

Also, tracing should be in the target, not an env var. This would let us
use the convention of ./bin/$(TARGET) for generator outputs instead of
repeating makefile rules for viz targets.