Use variadic templates to improve min/max for > 2 args
The real motivation here is that this allows a uniform syntax for
multiple -arg calls to min() and max(); the >2 arg version no longer
requires wrapper.

Additionally, by adding a variadic template function version of min()
and max(), we can preserve the proper behavior for arguments which are
ints (rather than Expr), which should be coerced to match the paired
Expr, rather than promoted to Expr first.


Former-commit-id: f6484ad57875d8508bfa33b06ac97d1acd2005a3