Don't use swap in our move constructors
The Buffer move constructors partially used swap (for alloc and
dev_ref_count) and partially used direct copying (for buf). This is
fishy in the case where the buffer being moved into already owns a
device allocation. The ownership may get swapped to the dying buffer but
the dev field may not.