Fix a bug that overwrote a part of the mattrs string on Hexagon.
This is sort of a latent bug; the reason this piece of code has worked
for so long is that when mcpu is hexagonv60 it has _always_ implied
hvx support. This may not always be true going forward.

The std::stringstream constructor puts the "put" pointer at the start
or 0. So, the very next call to "<<" will overwrite the contents that
the stream was constructed with. However, subsequent calls to "<<" will
work fine.