Aggressively prune specialize() branches
When a specialize() condition is const-false, we should always prune it
away before lowering.

When a specialize() condition is const-true, we should always prune
away all subsequent specializations (except for those that are also
const-true).

(Note that this pruning already effectively happened later during
simplification, but pruning earlier removes a nontrivial amount of
work.)