Nuke in-runtime texture tracking
This makes the GL runtime more like the other runtimes. The device field
is the native handle, and all the metadata is in the .buffer.

If you call device_free, it frees it unconditionally. It's on the user
to not call halide_device_free if they don't want the texture freed.


Former-commit-id: b5f8571db6bf5d5037efeae0108bf82cd453c933