Compiler performance penny-pinching
Two changes:

1) Use 32 free bits in the IRNode to store the IRNodeType, so that it
can be gotten at by a load instead of having to call a virtual function

2) Change things to a style where any function that's going to make a
copy of an Expr takes it by value, and then does a std::move internally
at its last use. This avoids a bunch of atomic ops and conditional
branches in caller code in the case where you're passing in an rvalue.

With these changes, lowering local laplacian gets about 12% faster (1.5s
-> 1.33s). Most of the win is from change #1

These are being done in advance of a planned change to simplify the
simplifier to be more concise and use less stack space, so hopefully the
fact that this represents a partial reversion of #1810 won't bite us.