Tighten up some Generator assumptions
Previously, the Generator class sorta-assumed it could be created with
a GeneratorContext (thus lazy-initing value_tracker and externs_map,
and defaulting Target to host); in practice, this is never the case
(aside from some internal-testing code inside Generator itself), and
attempting to use a Generator in this way is untested and possibly
buggy. This removes code for lazy-initing and adds asserts that verify
that a Generator always is created with a GeneratorContext.


Former-commit-id: 2c45d594114c04031c192159ee021f9c215024db