Allow creating & using Generators directly
This PR adds the minimal API necessary to allow for using Generators in
the same C++ translation unit as subcomponents of another pipeline,
without requiring the use of Stubs. This PR is not expected to be
accepted as-is, but rather as a proof-of-concept for adding this
functionality.

The major additions are Generator<T>::apply, which creates the
Generator, sets the Inputs, and calls the generate()/schedule()
methods; for cases where GeneratorParams need to be set, you can
instead use Generator<T>::create followed by
Generator<T>::generate_with_inputs.

Added a simple test (inlined_generator) for testing and as an example;
as a demonstration of the usefulness of this approach, also converted
the demosaic code in apps/camera_pipe to be in the form of a local
Generator (based largely on code from abadams@)

The goal of this change is to make it easier to use Generators as an
encapsulation layer in places where it isn’t convenient to do so right
now; e.g., in camera_pipe, we’d previously have had to move the
demosaicing code to another source file and access it via a Stub, which
is overhead that sometimes prevents a Generator-based approach from
being entertained at all.

I especially want comments on the need-to-set-GeneratorParam code flow,
and (of course) the methods that need better names…

(Note also that this branch includes the changes in PR #2064)


Former-commit-id: 1126f3de0713c984be0b451b679801e8910e4b4b