Add testing for C++ AOT backend
The C++ AOT backend is almost completely untested. This attempts to add test/generators to test using this backend in addition to the normal one. Caveats:

-- Previously, attempting to use this backend for vectorized code failed at compile time; now, if any vectorized types are found, we fail at runtime. (This is really a short-term workaround to simplify build rules.)
-- Extended Makefile to attempt to build all the Generator tests this way, then add extensive blacklists for the many that won't yet compile/link/run correctly.

Note that I didn't even attempt to get this working for CMake; IMHO we should postpone that until (at least) the blacklist of AOTCPP tests goes away.