Add Bazel build rules
This PR builds on the work done elsewhere (e.g. github.com/halide/halide_bazel) and intends to provide a stable, official set of Bazel build rules for Halide clients. (It is *not* intended to provide a way to build Halide itself using Bazel.)

The design is such that the neecessary Bazel build rules are added to Halide distributions, such that downstream consumers can add Halide usage to their project by a few lines added to their WORKSPACE.

The build rules are modeled very closely on those used inside Google, and are expected to be stable and supported for the long-term future.

These rules should work well for Linux and OSX hosts using Bazel 0.5 or later; Windows hosts are almost-but-not-quite supported, with a handful of issues that are expected to be resolved in subsequent PRs.

(Note that the Windows distrib is currently built by a script specific to the buildbots; after this PR lands, the buildbot script will be updated separately to add the new files to the distrib.)


Former-commit-id: 6d223d1ecec9170fe12acb1236926f07c0994899