avoid ambiguity between std::min and Halide::min
Avoids problems when people do:

using std::min;
using Halide::min;

It looks dumb when I write it like that, but the using std::min can be
buried in a header dependency somewhere.

This only happens because the stl has a templated min. Fortunately it
requires both template arguments to match, so we only have ambiguities
in cases where we call min/max on matching types that can both be cast
to Expr.


Former-commit-id: b3a47f2cbdb44985e56b0092e54f19bfe6ac113d