Allocate Functions in groups that share a refcount
This avoids reference cycles that cause memory leaks in master.

I introduce a FunctionPtr, which may be a weak or a strong reference to
a particular Function in a group (using the refcount of the entire
group). The programmer must obey several rules, which are not currently
mechanically checked. They are:

1) All references within a group must be weak.
2) If there are any references to the group from outside, at least one
such reference must be strong.
3) There may be no reference cycles that span multiple groups.

Func::in() and recursive update definitions take care to stay within one group and use weak references, to enforce 1 and 3. Calls from one Func to another form a DAG, and are strong to enforce 2.

When the deep copy happens I put everything in one big group, because it
was the simplest way to implement the deep copy. Deep copy therefore
changes all references to weak to enforce rule 1. During lowering, the
environment holds the only strong references to any Functions. At the
end of lowering, before the environment is discarded, the resulting stmt
is made to have strong references, enforcing rule 2. Transfer of Exprs
from the Function dag to the lowered stmt is one-way, so rule 3 is also
respected.