Clean up cropped output buffers
When an extern stage has a distinct compute_at/store_at, the computation
only writes to a subregion of the full buffer. This is implemented by
making a cropped alias of the output buffer. Didn't work at all for GPU
outputs, because it didn't crop the GPU allocation (fixed by earlier
commits), and it leaked any GPU resources given to the crop (fixed by
this commit).


Former-commit-id: 325e72cb39721018d8f2094b53ed2cf533ea3bb1