Dynamic tracking of folded coordinates
The assertions produced by storage folding when the compiler can't
statically prove safety were huge. This was mostly because they had to
be immediately inside the loop body (because they predicted the
footprint of the next loop iteration with a simple substitute, so you
can't let the loop var get hidden in any lets). Putting them there
forces them to include a copy of all the relevant letstmts that come
between them and the actual produce node.

By dynamically tracking what logical coordinates are currently
represented in the folded buffer, we can produce simpler asserts and
place them further in, immediately outside of the produce/consume nodes.

Moving the asserts further in is essential for getting storage folding
working for extern stages - they have to come after the dynamic bounds
queries. There are TODOs in this PR which will be resolved in that work.

Note that the dynamic tracking is only done for the benefit of the
asserts. So if we can statically prove the fold is safe, we don't do
dynamic tracking. This makes the overhead of the tracking relatively
insignificant.

Tested by the existing coverage in tests/correctness/storage_folding,
and by local laplacian. For local laplacian it generates simpler IR, but
there was no measurable performance difference.


Former-commit-id: 0d15358eff41fa927778628de82cbe171cfade2b