Add "makelib.sh" utility script
Some Generator-related rules could be made simpler if we could reliably combine .a files; libtool does this, but is not reliably available on all systems. This adds a simple bash script that allows combining of an arbitrary list of .a and .o files into a single .a file, and modifies the Makefile to make use of it in a few places.

Also, drive-by fix to the clean_generators target.