Change sliding window to use likely_if_innermost
In practice this means that in line-buffered pipelines loops over y
don't get partitioned on the y axis, which has two key benefits:

- The warm-up code in the y axis now gets partitioned in x, making
boundary conditions cost less.
- smaller binaries
- faster compile times

On local_laplacian and HDR+ this saves a bunch of code size and compile
time with no measurable performance difference. On camera_pipe this
seems to be a minor performance win.


Former-commit-id: 2bd064afc4e8cdb350c3ef0a5a8283f268ebd122