Require the -g argument for Generators
Previously, we allowed you to omit the -g argument to a Generator if there was exactly one Generator registered. This PR changes that to require -g be specified in all cases. (This change will allow some simplification of the Generator registration code, which will, in turn allow simplifying linking of Generators in some build systems).

Note 1: Bazel/Blaze build rules have always specified an explicit Generator name

Note 2: the CMake changes look a little ugly and redundant, but the CMake rules for Generators will be completely overhauled soon.