Remove apps/templates
This was written to be a simple apps framework for Halide, initially implemented on iOS and OSX; it hasn't been part of the build/test cycle, is currently broken, and will be Really Really Broken if the proposed CMake rule changes land. Rather than attempt to fix it (or leave broken code in the tree), let's delete it entirely. (We can always resurrect it from Git history if we decide it is worth taking the time to get it working again.)