Add ability to handle ctors and DT_INIT in the dynamic section.
This patch contains the following changes.
1) src/Elf.cpp : Defines a symbol called __CTOR_END__ that points
to the end of the list of pointers to constructors in the .ctors section.
It also sets up the DT_INIT tag to point to the start of the .init.halide
section.
2) src/HexagonOffload.cpp: Prepends (uint32_t) 0 to the start of the .ctors
section. Constructors are called by iterating backwards from (__CTOR_END__ -4)
until we hit a NULL pointer (delimiter). This 0 serves this purpose of being a
delimiter.
3) src/Elf.h: Added member functions prepend_contents and fixup_relocations_offset
to Section.
4) src/runtime/hexagon_remote/dlib.cpp: Process the DT_INIT tag and call initialization
function in mmap_dlopen.
5) src/runtime/qurt_init_fini.cpp: Initialization function that calls constructors whose
pointers are stored in .ctors.

This is now working on 8996 & 8998.