Rework halide_benchmark.h to use an adaptive iteration technique
This is mimicking a (very) simplified version of the approach used by GoogleTest's benchmark facility: instead of requiring the caller to specify explicit samples-and-iterations, we set a target minimum runtime (0.5 seconds), start with one iteration, scaling up the iteration count based on actual timing observed. (There are also knobs to control the min and max iters for pathological cases).

This tends to make most microbenchmarks simpler and more reliable (no guesswork about decent values), and also makes the results more familiar to users who are familiar with GTest's benchmarking facility (especially when used via RunGen).

Note that the "old" benchmark function is still present (for now); RunGen and test/performance have been converted to "new" style as proof-of-concept. (If this PR is approved then other uses will be converted in subsequent pulls.)