Allow Input<> to be used for Generators with build()
Up to now, we have explicitly disallowed using the Input<> syntax for Generators that used build(), rather than generate()+schedule(); this was done in the (misguided) thought that it would encourage conversion of Generators to the newer syntax. As it turns out, it actually *discouraged* said conversion, as it didn't allow for incremental changes to be made.

This PR changes that:
- You now may freely use Input<> in any Generator, even if it uses build().
- However, you may not mix Param/ImageParam and Input<> in the same Generator.
- All Generators in Halide have been converted to use Input<> rather than Param/ImageParam.
- It is still legal to use Param/ImageParam in Generators; however, they are now marked as deprecated when used in this context.