JITGeneratorContext -> GeneratorTarget
"JITGeneratorContext" was a misleading name since it isn't used only for JIT. Renamed to "GeneratorTarget" since it's really used to specify a Target to a Generator.

Also: driveby optimization to move externs_map and value_tracker out of GeneratorBase and completely into GeneratorContext, since the redundant implementations were, well, redundant.