JITGeneratorContext -> GeneratorTarget
"JITGeneratorContext" was a misleading name since it isn't used only for JIT. Renamed to "GeneratorTarget" since it's really used to specify a Target to a Generator.

Also: driveby optimization to move externs_map and value_tracker out of GeneratorBase and completely into GeneratorContext, since the redundant implementations were, well, redundant.

Former-commit-id: 5018ff78570ae8c378e242e639ae7b0ddb10cff7 [formerly dba7e2e0cadccd61f1f427d42bc398a158b03e4f]
Former-commit-id: 9cf34c0ce40cb5892db78cf2b1369654464f5325