Add a static "apply()" method to Generator Stubs
With the advent of late-bound ScheduleParams, it's now often tractable to use a Stub in "fire-and-forget" mode: since ScheduleParams can be set at any point, for many use cases a Stub doesn't need to be retained by an enclosing Generator, meaning that using a Stub is a whole lot like a simple function call.

Note that this doesn't remove any existing Stub functionality (I'm not yet sure whether we can avoid all possible 'stateful' uses in complex usages), although that might be a future possiblity if this works well.

Also completely rewrote example_jittest.cpp to both exercise the new functionality, and make existing usage clearer.

(Note to reviewers: this is based on top of https://github.com/halide/Halide/pull/2404, which must land first.)