Better error messages for device_api to device interface failures.
This will help users recognize and fix argument errors around device
API and the target it is paired with. In particular, requesting a
default GPU without specifying one in the target is a common
error. This is currently only wired up to report errors when used
through Halide::Buffer, but that is the common case. (Use of
error_site is both to hep pin down where the error occured and to give
a parameter that preserves the current behavior of returning nullptr
as it is likely some code depends on that.)

Add tests for errors. (Couldn't figure out an easy way to test missing
the runtime so that is not covered.)

Tab cleanup in unrelated file (implicit_args.cpp)