Add .mat support
This is the matlab level 5 format. It's not what matlab currently writes
(unless you set a flag), but matlab (and octave, and scipy, and
imagestack, ...) will be able to read it for the forseeable future, so
it's a good format for dumping things.

It's a fairly simple format, and it has also been added to ImageStack
already, so we should be able to just deprecate .tmp at some point if we
merge this.

TODO: Once HalideTraceDump is in, add .mat as a format option