Allow Output<Buffer<>> = Func
This eases converting Generators that used helper functions to produce Funcs that are returned from build().

Note the caveat about names: I chose to issue a warning rather than make this an error. not sure if this is the right decision.

(I also briefly thought about modifying Func::in() to allow making a global-wrapper with an arbitrary name, but that seems like a bad idea...)


Former-commit-id: 9d34058939313076ef3a49c7add241a78cf3f484 [formerly 75bd2629ca8864a56d667fa8e4d426fcdacdffe4]
Former-commit-id: d2621b1e0fca6b0e2cc55906f78a5439e6531768