Allow Output<Buffer<>> = Func
This eases converting Generators that used helper functions to produce Funcs that are returned from build().

Note the caveat about names: I chose to issue a warning rather than make this an error. not sure if this is the right decision.

(I also briefly thought about modifying Func::in() to allow making a global-wrapper with an arbitrary name, but that seems like a bad idea...)