Remove the SFINAE cruft from Generator
Replace the flaky SFINAE nonsense for build()/generate()/schedule() with plain old virtual methods.

The one gotcha here is that build()-returning-Pipeline isn't supported anymore. That's not a problem, though, since using Output<> is a cleaner way to do the same thing, and no instances of build()-returning-Pipeline exist in Halide. (There are a handful downstream that I should be able to convert with trivial effort, but this probably shouldn't land until we're sure we won't break downstream users.)