Make Generator aware of AutoScheduler
This revises the way AutoSchedule is integrated into Generator, with the intent of making it easier to use and harder to misuse; these changes will also pave the way to using Stubs with the auto-scheduler in a seamless way (in a subsequent PR).

Relevant changes:
-- Just as all Generators now have an implicit GeneratorParam for Target, they now have one for 'auto_schedule' and 'machine_params'. Also like 'target', these GeneratorParams are always propagated into sub-generators automatically.
-- Generators should no longer call auto_schedule_outputs() themselves (and in fact, this call is removed); Halide will call this for you after the toplevel Generator has finished.
-- You can still see the auto-schedule output in the debug(3) output, but the now-preferred way to get the output is by adding 'schedule' to -e, which saves it to a file named ${GENNAME}.schedule (basically, analogous to saving .stmt, etc).
-- Note that the auto-schedule output now includes the Target and MachineParams for which it was produced as comments.

I updated tutorial lesson 21 as I thought appropriate but it definitely needs a careful eye, esp. regarding MachineParams.


Former-commit-id: d73961bda7d79942c5601579f4594a601b4f88f5