Add apps/linear_blur
This is meant to show how to use Generators and Stubs with auto-schedule. Note that it is known to be awkward, mainly because the current auto-schedule setup is known to not fit well into Generators+Stubs as written; a subsequent PR (https://github.com/halide/Halide/pull/2446) will enable revisions that make this much less terrible.