Add apps/linear_blur
This is meant to show how to use Generators and Stubs with auto-schedule. Note that it is known to be awkward, mainly because the current auto-schedule setup is known to not fit well into Generators+Stubs as written; a subsequent PR (https://github.com/halide/Halide/pull/2446) will enable revisions that make this much less terrible.


Former-commit-id: f1a2f81394ebaf0bddab327c397171dbda26f39a [formerly 71ba02688e472f28b76787afe24ccf5a540e7d80]
Former-commit-id: 0c5e9ce374d8c5e98ba37b667af9bc321e236305