Expose source code locations for Func definitions
Feature request from some Stanford hardware folks, in order to be able
to report better debug info from their backends. The task of actually
plumbing the source location from the Definition object down to their
backend is yet to be done. It can be accessed via the FunctionPtr on a
Call node. We should possibly preserve it in some other way after
storage flattening (Maybe on ProducerConsumer nodes?)

It may also be interesting to expose this to tracing or the profiler.


Former-commit-id: 1ada2011191e5979d3fe8aadc978578ca79a3297