Add GeneratorParam<LoopLevel>
This essentially behaves the same as ScheduleParam<LoopLevel>, with some minor differences in Stub usage. This is intended as a sketch to demonstrate how ScheduleParam<> could be eliminated entirely (see https://github.com/halide/Halide/issues/2552), and shouldn't land until discussion there is clear on whether this is desirable.

(Note that this is built on top of https://github.com/halide/Halide/pull/2556)