Add GeneratorParam<LoopLevel>
This essentially behaves the same as ScheduleParam<LoopLevel>, with some minor differences in Stub usage. This is intended as a sketch to demonstrate how ScheduleParam<> could be eliminated entirely (see https://github.com/halide/Halide/issues/2552), and shouldn't land until discussion there is clear on whether this is desirable.

(Note that this is built on top of https://github.com/halide/Halide/pull/2556)


Former-commit-id: c1ce5f07c65c871ad3d6ff408743b7b947e89542 [formerly 1b5516f0b9eacf73bc4abb238587e50e7b5d5d78]
Former-commit-id: c17ea2b2a316ee57931d6f646b28abb8d5a0a52b