Make standalone-app Makefile helper (v2)
Currently, the apps/ folders that use Make use apps/support/Makefile.inc as a helper, which exposes them to ~everything about the internals of Halide. This PR proposes to add a "halide.make" helper file, which is intended to provided a minimum of assistance for apps that want to rely on **only a Halide distrib folder**.

Similar to the existing halide.bzl and halide.cmake helpers:
-- It's designed to be usable for most purposes by setting just HALIDE_DISTRIB_DIR
-- It's copied into the Halide distrib/ folder we produce, so downstream consumers can rely on using it with future binary drops
-- It remembers system libs required by the version of LLVM that Halide was built with, so there's a lot less weird voodoo about including specific platform-dependent libraries
-- It works more-or-less seamlessly with both the shared and static library builds of libHalide

Of course, unlike the Bazel and CMake helpers, it doesn't attempt to provide elegant build rules for Generators, because, well, it's Make, and I'm not sure if elegant canned build rules of the sort we need are really possible in Make; that said, it predefines a few things to make life easier downstream.

Proof-of-concept usage is in apps/wavelet, which is (currently) the example for "standalone" usage in apps/.

If the general feel of this seems OK, I'd propose to gradually convert all of the relevant apps/ to use this, to minimize coupling between the apps/ and the internals of Halide.