hexagon-lldb: Add "./" to shared libraries names given to dlopenbuf
This fix makes it possible to use hexagon-lldb on target to debug
code in libhalide_kernels*.so & libhalide_shared_runtime.so.

By convention, to debug shared libraries (located via search path),
lldb needs the library names returned from the dynamic loader to
begin with "./" which it replaces with directories in the search
path.  This is handled correctly by libraries opened via dlopen but
not by dlopenbuf.

The workaround is to explicitly add a beginning "./" to the
dlopenbuf library names.  A bug has also been opened to correct the
behavior of dlopenbuf so this workaround will eventually not
be needed.