Kill CMake support for python_bindings/
At the moment, it's not used or tested, and requires a lot of fiddling to make work; the Makefile is the only version that's regularly tested. The CMake version may be useful if we support Windows python bindings in the future (which we should), but at the moment, it's nonfunctional and confusing. Remove it entirely; we'll resurrect it from git history when we need it in the future.

(While we're there, re-order the targets in Makefile so that halide.so is the default target.)