Allow Param<void>
Disclaimer: the goal of this is really to simplify the Python Binding implementation, so it may be misguided. That said, the change is pretty modest and safe, I think.

The problem: in the Python bindings, there is no equivalent to a compile-time type bindng; for Param, the type is always specified by a Halide::Type, rather than a compile-time type.

Since the existing Param structure *requires* a compile-time type, the existing Python bindings jump through great hoops: it provides a set of bindings for every likely instantiation of Param<T>, then provides an extremely complex set of factory methods to create the right underlying wrapper based on the Type (and other arguments) you pass.

Since Param<> already contains dynamic type information (by way of Parameter::type()), this just expands on it (much as Buffer<void> vs Buffer<int> does): Param<void> allows the type to be specified at runtime (rather than compiletime), then does typechecking on the runtime type.

The one exception, of course, is the get() and set() method, which still need to be special-cased; however, this is *much* less onerous.

Realistically, no 'real' Halide code is ever likely to make use of Param<void>, but it does make binding to dynamic languages like Python substantially simpler.

(Note that Param<>::address() was totally unused, and so it was removed.)