Allow Buffer::copy() to work to/from Buffer<void>
make_with_shape_of() required the src and dst to be strongly typed, but there's no reason to do so; being able to copy to/from a Buffer<void> is fine as long as runtime types match.

Also, did unrelated drive-by fix in Bounds.h to fix an int-vs-size_t compiler warning for some configurations.