Add a way to manage lifetimes of non-owned Buffer data
Code that needs to use externally-allocated memory with Buffer (e.g., the Python bindings) needs a way to ensure that the Buffer can release the memory when all the users are done with it. This adds a quiet, backwards-compatible way to augment all the Buffer ctors that take explicit data fields, to allow specifying an optional keep-alive object.