Add ImageParam, OutputImageParam to py bindings
Added OutputImageParam and Dimension and finished ImageParam; added limited (but incomplete) testing.

Note the change to Dimension in the core code: the issue here is that in an attempt to prevent a const-Dimension from being copied into a non-const version and later mutated, we made the copy ctor private with friend access. This is untenable for the Python bindings; machinery inside PyBind11 needs to be able to copy/move them, and declaring 'friend' access would be a nightmare. Instead, make it public, and add the constness checks via a runtime flag.