Allow GeneratorParam<LoopLevel> to support assignment as well as mutation-in-place
We previously had a way to mutate-in-place (via set()), but we also need a way to support assignment (for stub usage to work well). This PR renames the existing suite of GeneratorParam 'set' related methods to be 'init', and enforces that they can only be called prior to the Generator's build()/generate() method; it also adds a set() method, specific to GeneratorParam<LoopLevel>, which continutes to offer the mutate-in-place behavior.