Rewrite CMake builds `install`/`distrib` targets.
This change adds support for a proper `install` target for CMake builds. This allows for the usage of CPack to create packages for various systems and platforms. Once Halide is ready to do these things, the settings can be adjusted accordingly. For now, the focus is on creating tgz and zip packages for POSIX systems and Windows, respectively. This is done using CPack.

The `distrib` target that was previously available has been rewritten in terms of the `install` target and taking advantage of being able to call CPack through CMake. As was done previously, a Halide.tgz package is created for POSIX systems while a Halide.zip package is created for Windows systems.

For Windows builds specifically, the following changes were made:
* Enable support for no RTTI builds
* Enable support for OpenMP builds
* Enable building of all but one app in the `apps/` folder
* Fix emitted code for the C codegen to not result in an "inconsistent dll linkage" warning

Builds were confirmed to be building with VS2017 on Windows.