Allow python generator stubs to work with "partial new-style" Generators
It turns out to be trivial to allow the Python stub bindings to support Generators that use Input<> (instead of ImageParam/Param), but still use build() (rather than Output<>+generate()). So, let's support that. Added tests to verified that new and partial-new Generators work in Python, and that old-style ones fail quickly and predictably.