Add stub support for synthetic GeneratorParams
For things like Input<Buffer<>> {"inputname"} (where there is no static type), we synthesize GeneratorParams of the form inputname.type (or inputname.dim, or inputname.size, in the case of buffer-dimensions or arraysize). These were settable from the build system, but not from stub usage (in either C++ or Python); this PR remedies that deficit and puts allows Stubs to set these. (Ditto for Outputs.)

Note that the syntax of "name.type" (using a dot) prevents us from using the same name in C++ structs or in Python named-arg calls; since a double-underscore is illegal in GeneratorParam names, we use that as a substitute in these contexts. (Note to reviewers: I'm not entirely happy with this nomenclature, and open for better suggestions.)